requirements.txt
Django==2.0.5
djangorestframework==3.8.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
websocket-client==0.48.0
channels==2.1.2
channels-redis==2.2.1
redis==2.10.6
aioredis==1.1.0
asgiref==2.3.2

pytest-django==3.3.3
pytest-asyncio==0.9.0

selenium==3.14.0
django-positions==0.6.0
